pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "test_task"
requires-python = "==3.11.*"
version = "1.0.0"
dependencies = [
    # For rest api
    "fastapi",
    "uvicorn",
    "jinja2",
    "python-multipart",
    # For database
    "chromadb",
    # Retrieve text from pdf
    "PyMuPDF",
    # Agents and LLMs
    "langchain",
    "langchain-huggingface",
    "langchain-community",
    "langchain-chroma",
    "sentence-transformers",
]

[project.optional-dependencies]
dev = [
    "pandas",
    "ipykernel",
    "ipywidgets",
    "matplotlib",
    "plotly",
    "grpcio==1.71.0",
    "ruff==0.6.4",
    "mypy==1.13.0",
    "pyright==1.1.389",
    "pre-commit==3.8.0",
]